working code to catch the pokemon
the code will verify the pokemon is still in the cache and attempt to catch if cache is available. It provides messages if the pokemon is no longer there, if the user fails to specify correct name when catching, and a success message.

for the purpose of testing i bumped up how long pokemon stays cached