Implement bare bones catch handler (#23)
* Just saving some of my stub work

* Now we are checking for catch message

* Add cache for messages at top level

Create a bot channel cache object and required configuration and pass it to the spawner. Is not yet passed to the primary handler.

* Make a global metadata variable for bot info

* working code to catch the pokemon

the code will verify the pokemon is still in the cache and attempt to catch if cache is available. It provides messages if the pokemon is no longer there, if the user fails to specify correct name when catching, and a success message.

for the purpose of testing i bumped up how long pokemon stays cached

* Code cleanup

* PR comment fixes and updated help since we have commands now

* Fix debounce window to be minimum spawn duration